backend/requirements.txt
fastapi[standard]==0.135.1
sqlmodel==0.0.37
python-jose[cryptography]==3.5.0